% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_ippo.R
\name{collate_ippo}
\alias{collate_ippo}
\title{Collate AAGI-CU IPPO register Excel workbooks for reporting}
\usage{
collate_ippo(dir_path_in, dir_path_out)
}
\arguments{
\item{dir_path_in}{A string value that provides the path to the top-level
directory of the R: drive holding the AAGI-CU Service and Support project
files.}

\item{dir_path_out}{A string value that provides the path to the desired
directory to write the output file to. \emph{It will be created if it does
not exist}. The file will be named "AAGI-CU-IPPO Register.docx" by default.
Any currently existing IPPO register will be overwritten.}
}
\description{
Traverses directories and imports IPPO register Excel(TM) workbook files and
collates them into a single unified reporting format.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 # for macOS
 library(fs)
 R_drive <- "/Volumes/0-9/AAGI_CCDM_CBADA-GIBBEM-SE21982/",
 collate_ippo(
   dir_path_in = path(R_drive, "Projects"),
   dir_path_out = path(R_drive, "Reports")
 )

 @returns Nothing, called for its side-effects of generating an MS Word(TM)
  Document of \acronym{IPPO} details for the purposes of \acronym{AAGI}
  reporting.
\dontshow{\}) # examplesIf}
}
