% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ippo_report.R
\name{create_ippo_report}
\alias{create_ippo_report}
\title{Create an IPPO register report for GRDC}
\usage{
create_ippo_report(tables_list, sp, outfile, infile)
}
\arguments{
\item{tables_list}{A nested list of IPPO register data tables for
\acronym{AAGI} service and support and research and development projects for
reporting generated with \code{list_ippo_tables()}.}

\item{sp}{A character string providing the AAGI strategic partner that is
generating the report, can be full name or two letter string used by
\acronym{AAGI}.}

\item{outfile}{A filename including the file path where the file is to be
written. The output file is an MS Word(TM) document that follows
\acronym{AAGI} style guidelines.}

\item{infile}{An optional filename providing the file path to an existing
IPPO report is to be read from. If not provided, a default empty template is
used. This can be used to add project IPPO details to an existing
\acronym{AAGI} IPPO register that has details that are already captured but
does not include service and support, research and development project IP.}
}
\value{
An invisible \code{NULL}, called for its side-effects of generating an
MS Word(TM) document with tables that report each project's IPPO.
}
\description{
Create an IPPO register report for GRDC
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

 # for macOS
 library(fs)
 sp <- "CU" # strategic partner for report
 R_drive <- "/Volumes/dmp/A-J/AAGI_CCDM_CBADA-GIBBEM-SE21982/"
 tl <- list_ippo_tables(
   dir_path_in = path(R_drive, "Projects"),
   sp = sp
 )
create_ippo_report(tables_list = tl,
                   outfile = "~/AAGI-CU-IPPO-register.docx",
                   sp = sp)
\dontshow{\}) # examplesIf}
}
\seealso{
Other reporting: 
\code{\link{list_ippo_tables}()}
}
\concept{reporting}
