% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ippo_tables.R
\name{create_ippo_tables}
\alias{create_ippo_tables}
\title{Create IPPO Tables List}
\usage{
create_ippo_tables(dir_path_in, dir_path_out)
}
\arguments{
\item{dir_path_in}{A string value that provides the path to the top-level
directory of the R: drive holding the AAGI-CU Service and Support project
files.}

\item{dir_path_out}{A string value that provides the path to the desired
directory to write the output file to. \emph{It will be created if it does
not exist}. The file will be named "AAGI-CU-IPPO Register.docx" by default.
Any currently existing IPPO register will be overwritten.}
}
\value{
A \code{list} object that contains five \code{\link[tibble:tibble]{tibble::tibble()}} objects.
}
\description{
Traverses directories and imports IPPO register Excel(TM) workbook files and
creates a list of tables corresponding to Tables 1-5 in the \acronym{AAGI}
\acronym{IPPO} register for all AAGI-CU Service and Support projects.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 # for macOS
 library(fs)
 R_drive <- "/Volumes/0-9/AAGI_CCDM_CBADA-GIBBEM-SE21982/"
 collate_ippo(
   dir_path_in = path(R_drive, "Projects"),
   dir_path_out = path(R_drive, "Reports")
 )
\dontshow{\}) # examplesIf}
}
