% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ippo_tables.R
\name{.filter_list}
\alias{.filter_list}
\title{Filter a list so that it only keeps objects that have a value}
\usage{
.filter_list(x)
}
\arguments{
\item{A}{list of objects to be filtered.}
}
\description{
Filter a list so that it only keeps objects that have a value
}
\keyword{internal}
