% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ippo_tables.R
\name{list_ippo_tables}
\alias{list_ippo_tables}
\title{Create IPPO tables list}
\usage{
list_ippo_tables(dir_path_in, sp)
}
\arguments{
\item{dir_path_in}{A string value that provides the path to the top-level
directory of the R: drive holding the AAGI-CU Service and Support project
files.}

\item{sp}{A character string providing the AAGI strategic partner that is
generating the report, can be full name or two letter string used by
\acronym{AAGI}.}
}
\value{
A \code{list} object that contains tables of IPPO registers organised by
AAGI Service and Support project or AAGI R&D Activity.
}
\description{
Traverses directories and imports IPPO register Excel(TM) workbook files and
creates a list of tables corresponding to Tables 1-5 in the \acronym{AAGI}
\acronym{IPPO} register for all AAGI-CU Service and Support projects.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 # for macOS
 R_drive <- "/Volumes/dmp/A-J/AAGI_CCDM_CBADA-GIBBEM-SE21982/"
 list_ippo_tables(
   dir_path_in = fs::path(R_drive, "Projects"),
   sp = "CU"
 )
\dontshow{\}) # examplesIf}
}
\seealso{
Other reporting:
\code{\link[=create_ippo_report]{create_ippo_report()}}
}
\concept{reporting}
