% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rippo-package.R
\docType{package}
\name{rippo-package}
\alias{rippo}
\alias{rippo-package}
\title{rippo: Simplify AAGI-CU IPPO Register Updates}
\description{
Helper functions that ease preparing the AAGI IPPO register from creating third party background IP tables to generating IPPO reports on a project by project basis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AAGI-AUS/rippo}
  \item \url{https://aagi-aus.github.io/rippo/}
  \item Report bugs at \url{https://github.com/AAGI-AUS/rippo/issues}
}

}
\keyword{internal}
