% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3.R
\name{create_table_3}
\alias{create_table_3}
\title{Create a Table of Third Party IP (Table 3) of the R packages and AAGI Quarto template used for an AAGI IPPO document}
\usage{
create_table_3(
  project_path = getwd(),
  outfile = "1 documentation/table_3.csv",
  aagi_project_code = "aagi-all-sp-003",
  quarto = false,
  digger = false
)
}
\arguments{
\item{project_path}{The path to the directory of the project to generate the
tables for, defaults to the current working directory and recurses into
subfolders.}

\item{outfile}{Character string that provides the file name and full path to
the directory where the table will be saved. Defaults to the current
project's \dQuote{1 Documentation} directory with a filename of
\dQuote{table_3.csv}.}

\item{aagi_project_code}{The \acronym{AAGI} project code to be inserted into
the table.  Defaults to the current Service and Support project code within
the \acronym{AAGI} project.  Note that this is NOT the project being
serviced but the code within the \acronym{AAGI} project that is allocated
for Service and Support, \emph{e.g.} "AAGI-ALL-SP-003".}

\item{quarto}{\code{Boolean} cite the Quarto software? Defaults to \code{FALSE}.}

\item{digger}{\code{Boolean} cite the \pkg{DiGGer} package? Defaults to \code{FALSE}.}

\item{asreml}{\code{Boolean} cite the \pkg{ASReml-R} package? Defaults to
\code{FALSE}.}
}
\value{
An invisible \code{NULL}, called for its side effects of writing a sheet
into an Excel workbook file (the project's IPPO register) on the local disk.
}
\description{
Automatically recognises \R packages that are used in a project and generates
Table 3 of the \acronym{IPPO} register.  Has allowances for
non-\acronym{CRAN} packages, \pkg{ASReml-R} and \pkg{DiGGer} and also the
Quarto software.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# after opening an RStudio project for a Service and Support Analysis

create_table_3()
\dontshow{\}) # examplesIf}
}
\seealso{
Other create IPPO: 
\code{\link{run_IPPO_app}()}
}
\author{
Zhanglong Cao, \email{zhanglong.cao@curtin.edu.au} and Adam H.
Sparks, \email{adam.sparks@curtin.edu.au}
}
\concept{create IPPO}
