% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rippo-package.R
\docType{package}
\name{rippo-package}
\alias{rippo}
\alias{rippo-package}
\title{rippo: Simplify AAGI-CU IPPO Register Updates}
\description{
Helper functions that ease preparing the AAGI-CU IPPO register.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AAGI-AUS/rippo}
  \item Report bugs at \url{https://github.com/AAGI-AUS/rippo/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adam.sparks@curtin.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Zhanglong Cao \email{zhanglong.cao@curtin.edu.au} (\href{https://orcid.org/0000-0001-6667-9392}{ORCID}) (Wrote original `create_table_3()`.)
  \item Braden Thorne \email{braden.thorne@curtin.edu.au} (\href{https://orcid.org/0000-0003-0497-1675}{ORCID}) (Contributed the Shiny app for generating IPPO register tables.)
}

Other contributors:
\itemize{
  \item Curtin University of Technology (Provided support through Adam Sparks, Zhanglon Cao and Braden Thorne's time.) [copyright holder]
  \item Grains Research and Development Corporation (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
