% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_IPPO_app.R
\name{run_IPPO_app}
\alias{run_IPPO_app}
\title{Run the Shiny IPPO app}
\usage{
run_IPPO_app()
}
\value{
An invisible \code{NULL}, called for its side-effects of launching a
Shiny application.
}
\description{
Runs a Shiny app where a user can build, edit and download an IPPO registry.
}
\examples{
\dontrun{
run_IPPO_app()
}
}
\seealso{
Other create IPPO: 
\code{\link{create_table_3}()}
}
\concept{create IPPO}
